+++ chaosazure/__init__.py
@@ -5,6 +5,7 @@
 from typing import List
 
 from azure.mgmt.compute import ComputeManagementClient
+from azure.mgmt.web import WebSiteManagementClient
 from azure.mgmt.resourcegraph import ResourceGraphClient
 from chaoslib.discovery import (discover_actions, discover_probes,
                                 initialize_discovery_result)
@@ -16,8 +17,8 @@ from chaosazure.auth import auth
 from chaosazure.common.config import load_configuration, load_secrets
 
 __all__ = [
-    "discover", "__version__", "init_client",
-    "init_resource_graph_client"
+    "discover", "__version__", "init_compute_management_client",
+    "init_website_management_client", "init_resource_graph_client"
 ]
 __version__ = '0.8.3'
 
@@ -34,10 +35,13 @@ def discover(discover_system: bool = True) -> Discovery:
     return discovery
 
 
-def init_client(
+def init_compute_management_client(
         experiment_secrets: Secrets,
         experiment_configuration: Configuration) -> ComputeManagementClient:
-
+    """
+    Initializes Compute management client for virtual machine,
+    and virtual machine scale sets resources under Azure Resource manager.
+    """
     secrets = load_secrets(experiment_secrets)
     configuration = load_configuration(experiment_configuration)
     with auth(secrets) as authentication:
@@ -50,9 +54,30 @@ def init_client(
         return client
 
 
+def init_website_management_client(
+        experiment_secrets: Secrets,
+        experiment_configuration: Configuration) -> WebSiteManagementClient:
+    """
+    Initializes Website management client for webapp resource under Azure
+    Resource manager.
+    """
+    secrets = load_secrets(experiment_secrets)
+    configuration = load_configuration(experiment_configuration)
+    with auth(secrets) as authentication:
+        base_url = secrets.get('cloud').endpoints.resource_manager
+        client = WebSiteManagementClient(
+            credentials=authentication,
+            subscription_id=configuration.get('subscription_id'),
+            base_url=base_url)
+
+        return client
+
+
 def init_resource_graph_client(
         experiment_secrets: Secrets) -> ResourceGraphClient:
-
+    """
+    Initializes Resource Graph client.
+    """
     secrets = load_secrets(experiment_secrets)
     with auth(secrets) as authentication:
         base_url = secrets.get('cloud').endpoints.resource_manager