+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 ## [Unreleased][]
 [Unreleased]: https://github.com/chaostoolkit-incubator/chaostoolkit-azure/compare/0.8.3...HEAD
 
+### Added
+
+-   Individual Azure managemnt clients for website and compute resources
+
+### Removed
+
+-   Removed common init_client
+
 ## [0.8.3][] - 2020-05-14
 
 [0.8.3]: https://github.com/chaostoolkit-incubator/chaostoolkit-azure/compare/0.8.2...0.8.3